Use size_t for lengths and indices.

+++ automata.c
@@ -7,7 +7,7 @@
 #define CELLS_WIDTH 500
 #define GENERATIONS 250
 
-bool pattern_matches(struct cells_t *cells, int idx, cell_state a, cell_state b,
+bool pattern_matches(struct cells_t *cells, size_t idx, cell_state a, cell_state b,
     cell_state c) {
 
   // TODO: Range checks
@@ -16,7 +16,7 @@ bool pattern_matches(struct cells_t *cells, int idx, cell_state a, cell_state b,
       && cells_get_state(cells, idx + 1) == c;
 }
 
-int rule_30_pattern(struct cells_t *cells, int i) {
+int rule_30_pattern(struct cells_t *cells, size_t i) {
     if(pattern_matches(cells, i, ON, OFF, OFF)) {
       return ON;
     } else if(pattern_matches(cells, i, OFF, ON, ON)) {
@@ -33,7 +33,7 @@ int rule_30_pattern(struct cells_t *cells, int i) {
 struct cells_t *rule_30(struct cells_t *cells) {
   struct cells_t *new_cells = cells_copy(cells);
 
-  for(int i = 1; i < cells_length(cells) - 1; i++) {
+  for(size_t i = 1; i < cells_length(cells) - 1; i++) {
     int newval = rule_30_pattern(cells, i);
     cells_set_state(new_cells, i, newval);
   }
